# services/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SOME/IP Simulation Studio",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/simulations": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Project Simulations",
        "operationId": "getProjectSimulations",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationWorkspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "Start Simulation",
        "operationId": "startProjectSimulation",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/runs": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Project Runs",
        "operationId": "getProjectRuns",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response Getprojectruns"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Run",
        "operationId": "getRun",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/stop": {
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "Stop Run",
        "operationId": "stopRun",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/inspection": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Inspection",
        "operationId": "getRunInspection",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunInspection"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{run_id}/captures/{capture_name}": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Download Capture",
        "operationId": "downloadRunCapture",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "capture_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Capture Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid project or workflow state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Project, document, run, or job was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The requested project resource already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ApiError"
      },
      "ArtifactFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          }
        },
        "type": "object",
        "required": [
          "name",
          "bytes"
        ],
        "title": "ArtifactFile"
      },
...
      "CaptureFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          },
          "wireshark_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Path"
          }
        },
        "type": "object",
        "required": [
          "name",
          "bytes"
        ],
        "title": "CaptureFile"
      },
      "CaptureSession": {
        "properties": {
          "wireshark_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Url"
          },
          "wireshark_root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Root"
          },
          "display_filter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Filter"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/CaptureFile"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "title": "CaptureSession"
      },
      "ContainerState": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "ContainerState"
      },
...
      "EvidenceState": {
        "properties": {
          "capture_status": {
            "type": "string",
            "title": "Capture Status"
          },
          "capture_count": {
            "type": "integer",
            "title": "Capture Count"
          },
          "wireshark_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wireshark Url"
          },
          "packet_snapshot_saved": {
            "type": "boolean",
            "title": "Packet Snapshot Saved"
          },
          "someip_observation": {
            "type": "string",
            "enum": [
              "captured",
              "unknown",
              "unavailable"
            ],
            "title": "Someip Observation"
          },
          "service_discovery_observation": {
            "type": "string",
            "enum": [
              "captured",
              "unknown",
              "unavailable"
            ],
            "title": "Service Discovery Observation"
          }
        },
        "type": "object",
        "required": [
          "capture_status",
          "capture_count",
          "packet_snapshot_saved",
          "someip_observation",
          "service_discovery_observation"
        ],
        "title": "EvidenceState"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NodeLifecycle": {
        "properties": {
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "status"
        ],
        "title": "NodeLifecycle"
      },
...
      "RunArtifacts": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ArtifactFile"
            },
            "type": "array",
            "title": "Files"
          },
          "capture": {
            "$ref": "#/components/schemas/CaptureSession"
          },
          "compose_log": {
            "type": "string",
            "title": "Compose Log",
            "default": ""
          },
          "capture_log": {
            "type": "string",
            "title": "Capture Log",
            "default": ""
          },
          "captures": {
            "items": {
              "$ref": "#/components/schemas/CaptureFile"
            },
            "type": "array",
            "title": "Captures"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunArtifacts"
      },
      "RunBlocker": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message"
        ],
        "title": "RunBlocker"
      },
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "capture": {
            "$ref": "#/components/schemas/CaptureSession"
          },
          "captures": {
            "items": {
              "$ref": "#/components/schemas/CaptureFile"
            },
            "type": "array",
            "title": "Captures"
          },
          "containers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ContainerState"
            },
            "type": "object",
            "title": "Containers"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "scenario_capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scenario Capabilities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "scenario_id",
          "scenario_name",
          "status"
        ],
        "title": "RunDetail"
      },
      "RunEvent": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "RunEvent"
      },
      "RunInspection": {
        "properties": {
          "run": {
            "$ref": "#/components/schemas/RunDetail"
          },
          "scenario_result": {
            "$ref": "#/components/schemas/ScenarioResult"
          },
          "node_lifecycle": {
            "items": {
              "$ref": "#/components/schemas/NodeLifecycle"
            },
            "type": "array",
            "title": "Node Lifecycle"
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidenceState"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/RunEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "artifacts": {
            "$ref": "#/components/schemas/RunArtifacts"
          }
        },
        "type": "object",
        "required": [
          "run",
          "scenario_result",
          "evidence",
          "artifacts"
        ],
        "title": "RunInspection"
      },
      "RunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "capture": {
            "$ref": "#/components/schemas/CaptureSession"
          },
          "captures": {
            "items": {
              "$ref": "#/components/schemas/CaptureFile"
            },
            "type": "array",
            "title": "Captures"
          },
          "containers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ContainerState"
            },
            "type": "object",
            "title": "Containers"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "scenario_id",
          "scenario_name",
          "status"
        ],
        "title": "RunSummary"
      },
...
      "Scenario": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "file": {
            "type": "string",
            "title": "File"
          },
          "steps": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "file"
        ],
        "title": "Scenario"
      },
      "ScenarioResult": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "status",
          "detail"
        ],
        "title": "ScenarioResult"
      },
      "SimulationStartRequest": {
        "properties": {
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          }
        },
        "type": "object",
        "required": [
          "scenario_id"
        ],
        "title": "SimulationStartRequest"
      },
      "SimulationWorkspace": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "scenarios": {
            "items": {
              "$ref": "#/components/schemas/Scenario"
            },
            "type": "array",
            "title": "Scenarios"
          },
          "blockers": {
            "items": {
              "$ref": "#/components/schemas/RunBlocker"
            },
            "type": "array",
            "title": "Blockers"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/TopologyNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Recent Runs"
          }
        },
        "type": "object",
        "required": [
          "project_id"
        ],
        "title": "SimulationWorkspace"
      },
...
      "TopologyNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "interface": {
            "type": "string",
            "title": "Interface"
          },
          "app_name": {
            "type": "string",
            "title": "App Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "interface",
          "app_name"
        ],
        "title": "TopologyNode"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}